code/requirements.txt
ray[rllib]
ray[tune]
https://s3-us-west-2.amazonaws.com/ray-wheels/latest/ray-0.8.0.dev6-cp37-cp37m-manylinux1_x86_64.whl
requests
setproctitle
psutil
tensorflow-gpu
tensorboard
plotly
git+https://github.com/vitchyr/viskit.git
aiohttp

